refactor(agent-core): 拆分 runtime 并整理 smoke
