+++ packages/agent-core/src/runtime/constants.ts
@@ -1,0 +1,6 @@
+export const TOOL_MEMORY_PREFIX = '[TOOL_MEMORY]';
+export const TOOL_MEMORY_RESULT_MAX_CHARS = 1200;
+export const TOOL_MEMORY_MAX_ENTRIES = 24;
+export const PENDING_WORKSPACE_FILE_MAX_CHARS = 24_000;
+export const STREAM_EVENT_BUDGET_PER_POLL = 128;
+export const STREAM_STALL_TIMEOUT_MS = 20_000;
